servers/queue/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
#mysql-connector-python==8.0.15
pika
shapely==1.5.13
pymongo
stomp.py
xmltodict
geopy